+++ CHANGELOG.md
@@ -4,7 +4,14 @@
 
 ## Features
 
-## Fixes
+- Implemented log context for log queries
+- Added configuration options for log context columns
+- Queries parsed from the SQL editor will now attempt to re-map columns into their correct fields for Log and Trace queries.
+- Added support for IN operator in adhoc filters
+
+### Fixes
+
+- Fixed and enhanced the logic for parsing a query back into the query builder.
 
 ## 4.4.0
 
@@ -12,7 +19,6 @@
 
 - Added "Labels" column selector to the log query builder
 - Datasource OTel configuration will now set default table names for logs and traces.
-- Added support for IN operator in adhoc filters
 
 ### Fixes